Search, automatic memberships, and downloader fixes
Added search functionality! Presents as an optional field on /get and /custom that searches the title field of posts in that subreddit.

Memberships are now automatically picked up through the role provided by the ko-fi discord bot.

Fixed bug with downloader truncating string in middle of character when string was past 256 characters and the last character is multiple bytes.