Error Handling, Async Redis
Errors are now propagated up to the command handler. If an error occurs, the command is responded to with an error code. The error is logged in posthog along with the shard_id and user. The error and code is then printed in stdout and a discord channel.

Redis now uses async connections. Hopefully faster.